check if .emacs.d exists before moving

+++ bin/bootstrap-workstation.sh
@@ -29,7 +29,9 @@ polite-git-checkout ~ https://github.com/joelmccracken/workstation.git
 
 cd ~
 
-mv ~/.emacs.d ~/.emacs.d-$(date +"%s")
+if test -e ~/.emacs.d; then
+  mv ~/.emacs.d ~/.emacs.d-$(date +"%s")
+fi
 git clone --depth 1 https://github.com/hlissner/doom-emacs ~/.emacs.d
 # install doom; `yes | ` answers yes for all prompts, however when doing this it
 # does not reliably exit, it just blocks forever even after the install finishes
